[Travis] Add clang to the build system
Since the default version of clang available in Travis doesn't support
C++17, a newer version has to be installed. The build the requires env
variables to be set accordingly (so that the latest version of clang is
used). Since similar steps are required for GCC, this patch creates a
Travis matrix to make this possible/easy.