Debug travis/Fix bug
The tests were failing on Travis (and only travis). The reason was that
I was using a comparison operator (<) for cci::List iterators. But the
comparison operator makes only sense for random access operators. In
particular, I can't assume that the nodes in a list are located in a
contiguous memory.